apps/science/requirements.txt
numpy
matplotlib
requests
requests_html
pandas
ta
PyYAML
